feat: added exceptions for product controller

+++ src/main/java/dev/shubham/productservice/services/DbProductService.java
@@ -1,12 +1,15 @@
 package dev.shubham.productservice.services;
 
 import dev.shubham.productservice.dtos.CreateProductRequestDto;
+import dev.shubham.productservice.exceptions.NoProductsException;
+import dev.shubham.productservice.exceptions.ProductNotFoundException;
 import dev.shubham.productservice.models.Category;
 import dev.shubham.productservice.models.Product;
 import dev.shubham.productservice.repositories.CategoryRepository;
 import dev.shubham.productservice.repositories.ProductRepository;
 import org.springframework.stereotype.Service;
 
+import java.util.Date;
 import java.util.List;
 import java.util.Optional;
 
@@ -22,13 +25,21 @@ public class DbProductService implements ProductService{
     }
 
     @Override
-    public Product getSingleProductDetails(Long id) {
-        return productRepository.findByIdIs(id);
+    public Product getSingleProductDetails(Long id) throws ProductNotFoundException {
+        Optional<Product> product = productRepository.findByIdIs(id);
+        if(product.isEmpty()){
+            throw new ProductNotFoundException();
+        }
+        return product.get();
     }
 
     @Override
-    public List<Product> getAllProducts() {
-        return productRepository.findAll();
+    public List<Product> getAllProducts() throws NoProductsException {
+        List<Product> products = productRepository.findAll();
+        if(products.isEmpty()){
+            throw new NoProductsException();
+        }
+        return products;
     }
 
     @Override
@@ -38,6 +49,7 @@ public class DbProductService implements ProductService{
         product.setDescription(createProductRequestDto.getDescription());
         product.setPrice(createProductRequestDto.getPrice());
         product.setImage(createProductRequestDto.getImage());
+        product.setUpdatedAt(new Date());
 
         Optional<Category> category = categoryRepository.findByTitle(createProductRequestDto.getCategory());
         if (category.isPresent()) {
@@ -52,16 +64,31 @@ public class DbProductService implements ProductService{
     }
 
     @Override
-    public Product getProductByTitleAndCategory(String title, String categoryTitle) {
-        return productRepository.findByTitleAndCategory(title, categoryTitle);
+    public Product getProductByTitleAndCategory(String title, String categoryTitle) throws ProductNotFoundException {
+        //TODO: Add exception to be thrown when more than 1 result is returned
+        Optional<Product> product = productRepository.findByTitleAndCategory(title, categoryTitle);
+        if(product.isEmpty()){
+            throw new ProductNotFoundException();
+        }
+        return product.get();
     }
 
     @Override
-    public Product getProductByTitle(String title) {
-        return productRepository.findByTitle(title);
+    public Product getProductByTitle(String title) throws ProductNotFoundException {
+        //TODO: Add exception to be thrown when more than 1 result is returned
+        Optional<Product> product = productRepository.findByTitle(title);
+        if(product.isEmpty()){
+            throw new ProductNotFoundException();
+        }
+        return product.get();
     }
+
     @Override
-    public void deleteProductById(Long id) {
+    public void deleteProductById(Long id) throws ProductNotFoundException {
+        Optional<Product> product = productRepository.findById(id);
+        if(product.isEmpty()){
+            throw new ProductNotFoundException();
+        }
         productRepository.deleteById(id);
     }
